requirements.txt
opencv_python==4.5.3.56
Pillow==8.3.1
